+++ causationentropy/core/discovery.py
@@ -1,7 +1,7 @@
 """
 Author: Kevin Slote
 Email: kslote@clarkson.edu
-version = 1.2.6
+version = 1.2.7
 """
 
 import copy